Added Dynamic types
It is now possible to expose a Rust type to the virtual machine and do
basic operations with it. Currently math operators aren't able to be
used with dynamic values, but that will be coming.

This change caused the VM to take a performance hit purely do to the
need for Drop to be called on Value. I can't think of any way around
this, but the fibonacci example still beats Python's version of the sae
code, so I'm still ok with the current performance.

In a way, it was unavoidable: to add strings or lists, I was going to
need Value to have a Drop implementation. By doing dynamic types before
the other types that need allocations, I open up the possibility that
the built-in types can be powered by associated types on the Envirnoment
trait, allowing users to swap out the String/List/Map types. Neat.