Refactored virtual machine for more speed.
These changes are inspired by me actually looking at how Lua is
designed. It shouldn't be too surprising to find out that a minimalistic
stack-based VM ends up being fairly similar, but I was a little
surprised to see how similar overall my design was to Lua's.

One major difference was that many operations, such as math operations,
don't actually act on the top of the stack. What I call variables are
called registers in Lua, and they're used throughout many operations to
avoid extra "noise" on the stack. While part of me thinks this is less
pure (and it definitely is more complex), the performance wins speak for
themselves.

Another change I made was adding the ability for the compare instruction
to jump, also inspired by Lua.

This more complex handling of values has caused the ast/langauge code to
become significantly slower due to generating a lot of unnecessary
temporary variables. This was something I was able to avoid in previous
code-gen, but the logic to determine the best location for a value is
too hard to track at the stage of code generation.

This means to regain the lost speed I'll need to start writing an optimizer
to clean up the temporary variable usage and removing extraneous clones.