Labels being removed are now accounted for
While reviewing the preview blog post, I noticed that the jump
instruction targeted an index that was 1 higher than it should. I
quickly realized it's because the IR language has an instruction for
labels, and the generate code has no actual instruction for the label.
Thus, the label's resolve instruction index needs to be offset by the
number of label instructions removed before the current one.