Added logic and bitwise operators
Originally my goal was to use the same operators, but I realized I
wanted short circuiting behavior on logic operations.

Ultimately, I chose to make the virtual machine instructions flexible
like I would expect a dynamic type system to do. However, this means
that a & b will downgrade to a logical or (producing a boolean) when
either a or b are not integers.

I'm not really sure which direction I want to take this. I may end up
changing it so that the bitwise operators try to convert the operands to
integers, but I don't have type conversions figured out yet.