Split Not into not/~
Closes #10. Closes #11.

This commit removes a lot of dead code (the overloaded operators for
logic weren't hooked up for Dynamic, but all those functions still
existed).

It also splits the VM Not instruction into BitwiseNot and LogicalNot.

Lastly, it adds a path for Dynamic values to allow themselves to be
coerced to an i64. This enables dynamic values to be used in
integer-only operations.