Reduced beef consumption, added new vm example
In the process of adding a new example showing how to interact with
Rust native functions and Dynamic types, I got tired of seeing Cow's all
over the examples. I couldn't get Into to work correctly with Cow due to
Cow<'a, [SomeType]> not implementing From for arrays. This meant that
&[Instruction::Call {..}] is actually &[Instruction; 1] which couldn't
be passed to a function expecting impl Into<Cow<'a, Instruction>>.

So, a new type that has more flexible From implementations was added.
This allowed me to remove Cows from all the examples.