Added Froms for Value, fixed convert edge case
For the Convert operation, the Kind wasn't being checked before
conversion was attempted. Now, if the kind is already what is being
requested, the value is just cloned and returned. Previously, attempting
to convert a String to a String would cause a fault due to the String's
DynamicValue implementation not supporting converting to itself.

The rest of this commit just adds some convenience From implementations
for Value.